Generate a stable service ID based on active configuration
Transport Canberra have entirely renumbered service IDs twice at this point (one week into the API being active). There is no indication that they won't recycle the number every time a service passes its end date.

Given this, a secondary unique and stable service ID needs to be generated. This is generated by hashing all the fields of the calendar.txt row that makes up a service. To be safe, the original service ID is also included in the hash, but this may not be necessary in the future and may be removed.